requirements.txt
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.24
streamlit
chromadb
python-dotenv
sqlalchemy
openai
pywin32
